+++ src/main/java/nextstep/presentation/front/ScheduleController.java
@@ -1,13 +1,10 @@
-package nextstep.presentation;
+package nextstep.presentation.front;
 
-import nextstep.domain.schedule.Schedule;
-import nextstep.dto.request.ScheduleRequest;
 import nextstep.dto.response.ScheduleResponse;
 import nextstep.service.ScheduleService;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
-import java.net.URI;
 import java.util.List;
 
 @RestController
@@ -18,21 +15,8 @@ public class ScheduleController {
         this.scheduleService = scheduleService;
     }
 
-    @PostMapping("/admin/schedules")
-    public ResponseEntity<Void> createSchedule(@RequestBody ScheduleRequest scheduleRequest) {
-        Long id = scheduleService.create(scheduleRequest);
-        return ResponseEntity.created(URI.create("/schedules/" + id)).build();
-    }
-
     @GetMapping("/schedules")
     public ResponseEntity<List<ScheduleResponse>> showReservations(@RequestParam Long themeId, @RequestParam String date) {
         return ResponseEntity.ok().body(scheduleService.findByThemeIdAndDate(themeId, date));
-    }
-
-    @DeleteMapping("/admin/schedules/{id}")
-    public ResponseEntity<Void> deleteReservation(@PathVariable Long id) {
-        scheduleService.deleteById(id);
-
-        return ResponseEntity.noContent().build();
     }
 }
